finish card and location name

+++ app/src/main/java/com/example/weathermood/model/Alert.kt
@@ -2,9 +2,9 @@ package com.example.weathermood.model
 
 data class Alert(
     val description: String,
-    val end: Int,
+    val end: Double,
     val event: String,
     val sender_name: String,
-    val start: Int,
+    val start: Double,
     val tags: List<String>
 )
