+++ app/src/main/java/com/example/weathermood/model/Current.kt
@@ -1,19 +1,19 @@
 package com.example.weathermood.model
 
 data class Current(
-    val clouds: Int,
+    val clouds: Double,
     val dew_point: Double,
     val dt: Long,
     val feels_like: Double,
-    val humidity: Int,
-    val pressure: Int,
+    val humidity: Double,
+    val pressure: Double,
     val rain: Rain,
-    val sunrise: Int,
-    val sunset: Int,
+    val sunrise: Long,
+    val sunset: Long,
     val temp: Double,
     val uvi: Double,
-    val visibility: Int,
+    val visibility: Double,
     val weather: List<Weather>,
-    val wind_deg: Int,
+    val wind_deg: Double,
     val wind_speed: Double
 )
