+++ app/src/main/java/com/example/weathermood/model/Minutely.kt
@@ -1,6 +1,6 @@
 package com.example.weathermood.model
 
 data class Minutely(
-    val dt: Int,
+    val dt: Double,
     val precipitation: Double
 )
